refactor: src/java/request 및 src/java/response 디렉토리 구조 변경
변경전: src/java/request, src/java/response
변경후: src/java/request/http, src/java/response/http: http 디렉토리 아래에 http 요청 및 응답 관련 클래스들을 포함시킴
