+++ src/main/java/controller/Controller.java
@@ -1,7 +1,7 @@
 package controller;
 
 import request.http.HttpRequest;
-import response.HttpResponse;
+import response.http.HttpResponse;
 
 public interface Controller {
     HttpResponse handleUserRequest(HttpRequest httpRequest) throws Exception;
