+++ src/main/java/response/http/StatusLine.java
@@ -1,4 +1,4 @@
-package response;
+package response.http;
 
 import util.StatusCode;
